requirements.txt
streamlit
camelot-py
pandas